+++ docs/CoreMetric.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / CoreMetric
+
+# CoreMetric
+More to come...
+
