+++ docs/IMetricTimingToken.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / IMetricTimingToken
+
+# IMetricTimingToken
+More to come...
+
