+++ docs/MetricTimingToken.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / MetricTimingToken
+
+# MetricTimingToken
+More to come...
+
