+++ docs/builders/CoreMetricBuilder.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / [Builders](/docs/builders/README.md) / CoreMetricBuilder
+
+# CoreMetricBuilder
+More to come...
+
