+++ docs/config/README.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / Configuration
+
+# Configuration
+More to come...
+
