+++ docs/exceptions/MetricConfigException.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / [Exceptions](/docs/exceptions/README.md) / MetricConfigException
+
+# MetricConfigException
+More to come...
+
