+++ docs/exceptions/README.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / Exceptions
+
+# Exceptions
+More to come...
+
