+++ docs/extensions/CoreMetricBuilderExtensions.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / [Extensions](/docs/extensions/README.md) / CoreMetricBuilderExtensions
+
+# CoreMetricBuilderExtensions
+More to come...
+
