+++ docs/extensions/README.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / Extensions
+
+# Extensions
+More to come...
+
