+++ docs/extensions/ServiceCollectionExtensions.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / [Extensions](/docs/extensions/README.md) / ServiceCollectionExtensions
+
+# ServiceCollectionExtensions
+More to come...
+
