+++ docs/outputs/ConsoleMetricOutput.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / [Outputs](/docs/outputs/README.md) / ConsoleMetricOutput
+
+# ConsoleMetricOutput
+More to come...
+
