+++ docs/outputs/IMetricOutput.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / [Outputs](/docs/outputs/README.md) / IMetricOutput
+
+# IMetricOutput
+More to come...
+
