+++ docs/outputs/README.md
@@ -1,0 +1,5 @@
+[Home](/README.md) / Outputs
+
+# Outputs
+More to come...
+
