+++ src/Rn.NetCore.Metrics/Builders/CoreMetricBuilder.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 
 namespace Rn.NetCore.Metrics.Builders;
 
+// DOCS: docs\builders\CoreMetricBuilder.md
 public interface ICoreMetricBuilder<TBuilder>
 {
   ICoreMetricBuilder<TBuilder> AddAction(Action<CoreMetric> action);
