+++ src/Rn.NetCore.Metrics/Configuration/RnMetricsConfig.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.Configuration;
 
 namespace Rn.NetCore.Metrics.Configuration;
 
-// DOCS: docs\configuration\appsettings.md
+// DOCS: docs\config\RnMetricsConfig.md
 public class RnMetricsConfig
 {
   public const string ConfigKey = "Rn.Metrics";
