+++ src/Rn.NetCore.Metrics/IMetricTimingToken.cs
@@ -1,0 +1,6 @@
+using System;
+
+namespace Rn.NetCore.Metrics;
+
+// DOCS: docs\IMetricTimingToken.md
+public interface IMetricTimingToken : IDisposable { }
