+++ src/Rn.NetCore.Metrics/Outputs/IMetricOutput.cs
@@ -1,8 +1,9 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Rn.NetCore.Metrics.Outputs;
 
+// DOCS: docs\outputs\IMetricOutput.md
 public interface IMetricOutput
 {
   bool Enabled { get; }
